Refactor bloom filter
